Отложено зареждане на макро графиката
